fixes: robot visual fk for mixed mesh frame convention
NOTES (IMPORTANT):
- Some robot assets (e.g. Z1 CAD-derived meshes) encode their link frames directly in mesh local space, while others (URDF-style, e.g. Panda) expect FK to be applied at the object level
- Engine-aligned models: bake FK directly into mesh local transform
- URDF-style models: apply FK + visual origin to object transform
- MAY REVISE LATER

+++ Sim_Engine/EngineLib/include/Scene/SimulationManager.h
@@ -155,7 +155,7 @@ namespace gui {
 
 		// Robot System
         void loadRobot(const std::string& name);
-        void setRobotLinkRotation(const std::string& linkName, float angle);
+        void setRobotLinkRotation(const std::string& linkName, double angle);
         void setRobotRootPose(const glm::vec3& pos, const glm::quat& rot);
         void setRobotRootHome(const glm::vec3& pos, const glm::quat& rot);
         void resetRobot();
